backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# MongoDB
mongoengine==0.27.0
djongo==1.3.6
pymongo==4.6.0

# Environment variables
python-decouple==3.8
python-dotenv==1.0.0

# Image handling
Pillow==10.1.0

# Data validation
marshmallow==3.20.1

# Password hashing
argon2-cffi==23.1.0

# API documentation
drf-yasg==1.21.7

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.12.0
isort==5.13.2

# Utilities
python-slugify==8.0.1
redis==5.0.1

# Development
django-debug-toolbar==4.2.0
ipython==8.18.1
